requirements.txt
pydantic-ai~=0.3.2
a2a-sdk~=0.2.12
allure-python-commons~=2.14.3
allure-combine~=1.0.11